+++ examples/AspNetCoreWebApp/ApplicationInsightsDiagnostics.json
