+++ examples/AspNetCoreWebApp/AspNetCoreWebApp.csproj
