+++ examples/AspNetCoreWebApp/Models/ErrorViewModel.cs
