+++ examples/AspNetCoreWebApp/Properties/launchSettings.json
