+++ examples/AspNetCoreWebApp/Views/Home/Privacy.cshtml
