+++ examples/AspNetCoreWebApp/Views/Shared/Error.cshtml
