+++ examples/AspNetCoreWebApp/Views/Shared/_Layout.cshtml
