+++ examples/AspNetCoreWebApp/Views/Shared/_ValidationScriptsPartial.cshtml
