+++ examples/AspNetCoreWebApp/Views/_ViewImports.cshtml
