+++ examples/AspNetCoreWebApp/Views/_ViewStart.cshtml
