+++ examples/AspNetCoreWebApp/appsettings.json
