+++ examples/BackgroundTasksWithHostedService/HostedServices/TimeHostedService.cs
@@ -2,9 +2,8 @@
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
