+++ examples/BackgroundTasksWithHostedService/Program.cs
@@ -1,11 +1,15 @@
-﻿using System.Threading.Tasks;
-using BackgroundTasksWithHostedService.HostedServices;
+﻿using BackgroundTasksWithHostedService.HostedServices;
+
+using Microsoft.ApplicationInsights.Channel;
+using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+
+using System.Threading.Tasks;
 
 namespace BackgroundTasksWithHostedService
 {
@@ -58,6 +62,32 @@ namespace BackgroundTasksWithHostedService
 
                 // Wait for the host to shutdown
                 await host.WaitForShutdownAsync();
+            }
+        }
+
+        internal class MyCustomTelemetryInitializer : ITelemetryInitializer
+        {
+            public void Initialize(ITelemetry telemetry)
+            {
+                // Replace with actual properties.
+                (telemetry as ISupportProperties).Properties["MyCustomKey"] = "MyCustomValue";
+            }
+        }
+
+        internal class MyCustomTelemetryProcessor : ITelemetryProcessor
+        {
+            ITelemetryProcessor next;
+
+            public MyCustomTelemetryProcessor(ITelemetryProcessor next)
+            {
+                this.next = next;
+            }
+
+            public void Process(ITelemetry item)
+            {
+                // Example processor - not filtering out anything.
+                // This should be replaced with actual logic.
+                this.next.Process(item);
             }
         }
     }
