+++ examples/ConsoleApp/ConsoleApp.csproj
