+++ examples/ConsoleApp/Program.cs
