+++ examples/Examples.sln
@@ -51,13 +51,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Web", "..\WEB\Src\Web\Web\W
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SDKs", "SDKs", "{C5024E50-3350-479B-8628-620B299B9061}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApp.AspNetCore", "WebApp.AspNetCore\WebApp.AspNetCore.csproj", "{5DB76934-58D9-45BD-91A9-E1D78F4AE17D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreWebApp", "AspNetCoreWebApp\AspNetCoreWebApp.csproj", "{5DB76934-58D9-45BD-91A9-E1D78F4AE17D}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BackgroundTasksWithHostedService", "BackgroundTasksWithHostedService\BackgroundTasksWithHostedService.csproj", "{9E35EFBE-F478-4DEB-9327-C712272B77CA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleAppWithApplicationInsights", "ConsoleAppWithApplicationInsights\ConsoleAppWithApplicationInsights.csproj", "{DD21C077-1345-459C-8704-9BCD6F312029}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp", "ConsoleApp\ConsoleApp.csproj", "{DD21C077-1345-459C-8704-9BCD6F312029}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorkerServiceSampleWithApplicationInsights", "WorkerServiceSampleWithApplicationInsights\WorkerServiceSampleWithApplicationInsights.csproj", "{3DA486F4-10E3-432D-9916-2C1EE15029C7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorkerService", "WorkerService\WorkerService.csproj", "{3DA486F4-10E3-432D-9916-2C1EE15029C7}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
