+++ examples/WorkerService/WorkerService.csproj
