--- examples/WorkerServiceSampleWithApplicationInsights/MyCustomTelemetryInitializer.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.ApplicationInsights.Channel;
-using Microsoft.ApplicationInsights.DataContracts;
-using Microsoft.ApplicationInsights.Extensibility;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace WorkerServiceSampleWithApplicationInsights
-{
-    public class MyCustomTelemetryInitializer : ITelemetryInitializer
-    {
-        public void Initialize(ITelemetry telemetry)
-        {
-            // Replace with actual properties.
-            (telemetry as ISupportProperties).Properties["MyCustomKey"] = "MyCustomValue";
-        }
-    }
-}
